resolve shuo's comments
